UI: Rename next server files to use - as word separator
