+++ monkey/monkey_island/cc/next_ui/generated-next-config.js
