+++ monkey/monkey_island/cc/next_ui/server-common.js
