+++ monkey/monkey_island/cc/setup/nextjs/nextjs_process.py
@@ -16,7 +16,7 @@ UI_DIR = os.path.join(MONKEY_ISLAND_ABS_PATH, "cc", "next_ui", ".next", "standal
 _NODE_EXECUTABLE_PATH_WIN = os.path.join(MONKEY_ISLAND_ABS_PATH, "bin", "node", "node.exe")
 _NODE_EXECUTABLE_PATH_LINUX = os.path.join(MONKEY_ISLAND_ABS_PATH, "bin", "node", "node")
 NODE_EXECUTABLE_PATH = _NODE_EXECUTABLE_PATH_WIN if is_windows_os() else _NODE_EXECUTABLE_PATH_LINUX
-NEXTJS_EXECUTION_COMMAND = [NODE_EXECUTABLE_PATH, "server_prod.js"]
+NEXTJS_EXECUTION_COMMAND = [NODE_EXECUTABLE_PATH, "server-prod.js"]
 
 
 class NextJsProcess:
